always listen for text events
before, they were disabled when in unpaused gameplay. the proper way to handle this would be to enable/disable it conditionally when a textbox is actually active and receiving input.

this should fix chat while playing, but will break IME input if anyone was using that (unless "use_ime" is set to true)